feat: move logic out of deployment ui
